Arreglo test ColeccionPalabra y agrupo clases en carpeta correspondiente
- Pequeña correccion coleccion palabras
- Agrupo Tests en carpeta correspondiente
